Split out MVMStrHashTableControl from MVMStrHashTable.
And similarly for the other 4 types of hash.

The "control" structure contains all the members that used to be in the main
structure; the main structure is now just a pointer to the control structure.
This commit introduces an extra level of indirection, but the next commits
will remove it.

It also causes a *lot* of code churn replacing `hashtable` with `control`.