Convert debug_locals from `MVMStrHashTable *` to `MVMStrHashTable`.
Now that the public hash table structure is itself just a single pointer,
for all the code that only sometimes wants to actually build a hash table,
there is now no space saving in having a pointer to a hash table.

We do need to take a bit more care about whether the hash table has been
"built", but the existing iterator logic is robust against handling an
unbuilt hash. (It behaves identically to a "built" hash with no entries.)